chore: simplify uopool routing
