+++ Cargo.toml
@@ -12,6 +12,7 @@ anyhow = "1"
 arrayref = "0.3"
 async-trait = "0.1"
 clap = { version = "4", features = ["derive"] }
+dashmap = "5.4.0"
 dirs = "4.0"
 educe = { version = "0.4", features = ["Debug", "Default"] }
 ethereum-interfaces = { git = "https://github.com/ledgerwatch/interfaces" }
