--- src/uopool/services/mod.rs
@@ -1,5 +1,0 @@
-mod sanity_check;
-mod simulation;
-mod uopool;
-
-pub use self::uopool::*;
